View layout & style tweaks
Added a helper to app.rb for creating nice page titles. Updated layout
and most views to use it.

Moved bootstrap container and footer to layout. Previously, pages
included both the div for the container and the footer. This allows the
footer to be updated without having to update individual views.

Major styling & layout tweaks to 'main_page' view. It should look
better.