Added '/api/login' route for non-browser clients
This route was previously removed because it had redundant auth code.
The auth code is now in a helper method and /login and /api/login call
it.

I'm currently testing admin automation using rest-client. User auth
currently uses cookies, and rest-client is unable to get the cookie w/
session info when POSTing to /login due to its redirect to /main.

/api/login does not redirect, so rest-client can get the cookie from its
response.

Browser clients will continue to use /login, as it's easier and
requires no JavaScript.