Added /api/payment route and 'rankup' payment type
The 'rankup' payment allows user to increase their rank up to a
(current) maximum of 5. In the future, the rank will be shown by their
names when going all-in.

models.rb needed some temporary adjustments to control rank number.

I also added a bit of debugging-friendly text to Payment validation.